Rename grid to module_grid
