+++ source/source_base/module_grid/test/test_partition.cpp
@@ -1,6 +1,6 @@
-#include "source_base/grid/partition.h"
-#include "source_base/grid/radial.h"
-#include "source_base/grid/delley.h"
+#include "source_base/module_grid/partition.h"
+#include "source_base/module_grid/radial.h"
+#include "source_base/module_grid/delley.h"
 #include "source_base/constants.h"
 
 #include "gtest/gtest.h"
