+++ source/source_base/module_grid/test/test_radial.cpp
@@ -1,4 +1,4 @@
-#include "source_base/grid/radial.h"
+#include "source_base/module_grid/radial.h"
 
 #include "gtest/gtest.h"
 #include <cmath>
